:recycle: :truck: :fire: Canvas only has required code
Will have more than one stego algorithm
(not just lsb) so created file structure, where
common stego functions will be and specific one to the algorithm will be in
their respective folders.

Canvas component only has the functions it
needs, so removed the buttons and also the
get and set functions for pixel data.

Steganography component which has common code for stego algorithms will
now take a webview as an argument and get and set the pixel data.


Former-commit-id: dbf06a3a9f89ce90b7a39f5403c8f29186f12f18 [formerly 05b3630c658697219a5ead5c919a73afa3636913]
Former-commit-id: 0422347b48446a0f537bfdf7660f1a039c2c46c2