:pencil2: :bug: :arrow_up: Upgraded to React Navigtion V3
Upgraded to v3 for react navigation didn't have to change too much
mostly wrap all the react navigation items in a container. Everything
else is pretty much the same.

Fixed a bug with theme not changing based on the page for FAQ or About
because the reducer function wasn't updated added a quinary colour
for the FAQ.