:sparkles: :fire: :robot: Added Automatic Theme
Added the feature automatic theme where the user can select this option
in the settings. During the day it will be a light theme and during the
night the dark theme automatically.

It checks the sunrise and sunset time using lat long coordinates. It
checks everytime the app is opened or back from the background. The lat
longs are only check every week. If we cannot get the lat long use the
last known.

If we cannot get lat. long turn off the automatic theme option for the
user.

Fixes #2