:children_crossing: Warn if message too long
Warns the user if the message they wan't to encode it too long.
Better UI/UX as know the user knows if they need to shorten their
message.

In `message.tsx` (encoding) we get the image size/number of pixels it has
multiple by 3/4 (we don't use Alpha channel). Then we use the `Steganography.ts`
class to see how many bits we need for encoding. If bits to encode > bits available
warn the user


Former-commit-id: 4754662a4492184500c3d82acea28325c2dce669