:truck: :art: :bug Moved /functions
Moved /functions to src/services. All Firebase fuctions will be in services. So currently it only contains steganography. This makes it much cleaner however and fits in with the project structure better. Rather than having a random /functions folder at the root level.

Also fixed a bugs with firebase deploy, tested it works and also hide stdout so we don't see the token. 11.10 should fix this issue, https://gitlab.com/gitlab-org/gitlab-ce/issues/13784. So we don't have to hide all the stdout we can just hide the token.


Former-commit-id: ae3463241364e69a4d49c25b4da6b71ce664af88