:bug: Fixed DrawerNavigator Bug
Fixed error where you could see DrawerNavigator in background.
The error turned out be related to this line `this.setState({ appState: nextAppState })`, in `MainApp.tsx`. Turns out
this was redundant code, don't need any state in that module. This was used to set the app state i.e. if the app is in focus or in the background. Removing this line seems to fix my issue with the DrawerNavigator.

Fixes #61


Former-commit-id: d3cd218a55234a0dc902a7e51af65b3e648873a6