:children_crossing: Added Share Icon
Added a share icon on the progress encoding page, makes it more clear
to click the progress circle to share the image. Home icon shows on top
right of app header, when on decoding/encoding pages except the main ones.
The home button then takes you to the main page for that section.