:white_check_mark: :art: Improved Test Data
Improved test data so we don't store the image data or base64 string within the `data.ts` file instead we include images (encoded and normal). Then we use functions in `data.ts` to convert them to a base64 string and image data. Also edited tests to do the encoding and decoding in one step as the encoded image data doesn't always match the base64 image, even though it decodes it correctly.