:fire: Removed F5 and JSTEG
Removed F5 and JSTEG algorithms from the application, only allow user to use LSB.
This was due to algorithms not being implemented properly not working or taking
too long. Made a few smaller changes such as the snackbar text and fixed a small
bug with the PhotoAlbumList not loading new photos.