:sparkles: :robot: Auto Publish to Play Store
Integrated the Gradle Play Publisher (GPP) into the CI/CD.
We can now automate publishing of everything to the play store,
including the store listing and release notes. At the moment
it auto publishes to the internal track, we still have to
manually deploy to production. But this is so that we manually
test the app first then release into production.