:bug: :rocket: Fixed Android Play Store Build
The app was getting stuck at the splash screen when
downloaded from the play store. This was fixed
by turning off hermes.

Also changed the publish package logic slightly
now dev -> alpha and internal is pushed to
by running the script locally. This means it's
a lot easier to test the app locally (from Play Store).